moved to RichStorageLevel and added test
